apps/tts-server/pyproject.toml
[project]
name = "tts-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coqui-tts>=0.27.2",
    "fastapi>=0.119.0",
    "torch>=2.8.0",
    "uvicorn>=0.37.0",
]
